backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
requests==2.31.0
psycopg2-binary==2.9.9
gunicorn==22.0.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
